// src/timesheet-web/src/app/api/models/backlog-update-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BacklogUpdateRequest {
  assigneeUserId?: number | null;
  auditNote?: string | null;
  backlogCode?: string | null;
  deadlineExternal?: string | null;
  deadlineInternal?: string | null;
  endDate?: string | null;
  expectedVersion?: number;
  note?: string | null;
  officialEstimateHours?: number | null;
  pcaContactId?: number | null;
  periodMonth?: string | null;
  progressPercent?: number | null;
  project?: string | null;
  roughEstimateHours?: number | null;
  startDate?: string | null;
  type?: string | null;
}